requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
pydantic>=2.0
pydantic[email]>=2.0
pydantic-settings
python-dotenv
argon2-cffi
passlib[bcrypt]
python-jose[cryptography]
pytest
pyjwt
faker
pytest-asyncio 
httpx
aiosqlite
